Changed price type to decimal

+++ init.sql
@@ -88,7 +88,7 @@ CREATE TABLE Orders
     Branch      int NOT NULL,
     Publisher   int NOT NULL,
     Bookstore   int,
-    totalPrice  int NOT NULL,
+    totalPrice  decimal NOT NULL,
     FOREIGN KEY (Customer) REFERENCES Customer (CustomerID),
     FOREIGN KEY (Branch) REFERENCES Branch (BranchID),
     FOREIGN KEY (Publisher) REFERENCES Publisher (PublisherID),
@@ -101,8 +101,8 @@ CREATE TABLE Books
     Subject      varchar(255),
     Author       int          NOT NULL,
     Title        varchar(255) NOT NULL,
-    SellingPrice int          NOT NULL,
-    CostPrice    int          NOT NULL,
+    SellingPrice decimal          NOT NULL,
+    CostPrice    decimal          NOT NULL,
     FOREIGN KEY (Author) REFERENCES Author (AuthorID)
 );
 
@@ -112,7 +112,7 @@ CREATE TABLE OrderDetails
     OrderNumber    	 int NOT NULL,
     ISBN           	 int NOT NULL,
     Quantity       	 int NOT NULL,
-    orderDetailPrice int NOT NULL,
+    orderDetailPrice decimal NOT NULL,
     FOREIGN KEY (OrderNumber) REFERENCES Orders (OrderNumber),
     FOREIGN KEY (ISBN) REFERENCES Books (ISBN)
 );
